+++ source/extensions/omni.isaac.tests/omni/isaac/tests/tests/__init__.py
@@ -7,4 +7,9 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 
-from .extension import *
+from .test_physics import *
+from .test_utility_snippets import *
+from .test_external import *
+from .test_jetbot import *
+from .test_jetracer import *
+from .test_carter_v2 import *